frontend/requirements.txt
Pillow==9.5.0
Requests==2.28.2
streamlit==1.21.0
